Update image URLs to use IPFS gateway (#433)
* migrate constants

* feat: Update image URLs to use IPFS gateway

This commit updates the image URLs in the PxWebinar, courseRelease, PxFooter, and PxHeader components to use the IPFS gateway. The image URLs are now constructed using the `gateway` variable from the `IPFS` constant file. This change ensures that the images are loaded from the correct IPFS gateway and improves the reliability and performance of the application.

* Update IPFS image URL in certificate.html
