+++ api/templates/certificate.html
@@ -103,7 +103,7 @@
             </div>
             <div></div>
             <div style="position:absolute; bottom:0; padding: 3rem 0rem;">
-                <img src="https://moccasin-perfect-trout-941.mypinata.cloud/ipfs/QmdttBPgdS8ERmJQgacHire1y2F8uHJLEm6oNmSx9yisbV"
+                <img src="https://blind-gallery.infura-ipfs.io/ipfs/QmdttBPgdS8ERmJQgacHire1y2F8uHJLEm6oNmSx9yisbV"
                     width="170px">
             </div>
         </div>
