+++ ui/components/landing/courseRelease.vue
@@ -1,3 +1,13 @@
+<script>
+import { IPFS } from '@/constants'
+export default {
+  data () {
+    return {
+      gateway: IPFS.GATEWAY
+    }
+  }
+}
+</script>
 <template>
   <div class="tw-container">
     <div class="tw-grid tw-grid-cols-1 lg:tw-grid-cols-2 tw-shadow-sm tw-rounded tw-my-8 hover:tw-shadow tw-duration-200 tw-ease-in-out  tw-border-l-4 tw-border-l-cyan-500">
@@ -22,10 +32,10 @@
         </NuxtLink>
       </div>
       <div class="tw-w-full tw-flex tw-items-center tw-justify-center lg:tw-justify-start  tw-order-1 lg:tw-order-2 tw-relative">
-        <img src="https://moccasin-perfect-trout-941.mypinata.cloud/ipfs/QmQ6xcW8YcPdCVxM82mAjc9qSTmyks3buzaxsynjQZndby" class="tw-w-full tw-h-auto" alt="Imagen">
+        <img :src="`${gateway}QmQ6xcW8YcPdCVxM82mAjc9qSTmyks3buzaxsynjQZndby`" class="tw-w-full tw-h-auto" alt="Imagen">
         <div class="tw-absolute  tw-inset-0 tw-bg-gradient-to-l tw-from-transparent tw-to-white tw-opacity-100" />
         <div class="tw-absolute tw-overflow-hidden tw-rounded-full md:tw-w-[170px] md:tw-h-[170px] tw-w-[150px] tw-h-[150px]">
-          <img src="https://moccasin-perfect-trout-941.mypinata.cloud/ipfs/QmV2PBdksQNvsyMfwxcXVvq4Y22BA8rXWSVmTFnMkhB9FG">
+          <img :src="`${gateway}QmV2PBdksQNvsyMfwxcXVvq4Y22BA8rXWSVmTFnMkhB9FG`">
         </div>
       </div>
     </div>
