+++ ui/constants/allowList.js
@@ -1,0 +1,1 @@
+export const EARLY_ACCESS_USER_IDS = process.env.EARLY_ACCESS_USER_IDS ?? []
