+++ ui/constants/ipfs.js
@@ -1,0 +1,1 @@
+export const GATEWAY = 'https://blind-gallery.infura-ipfs.io/ipfs/'
