+++ ui/constants/metadata.js
@@ -1,0 +1,1 @@
+export const APP_NAME = 'Academy by Blind Gallery'
